requirements.txt
azure-functions
beautifulsoup4
requests
python-docx
pdfreader
filetype